backend/requirements.txt
flask
flask-security-too
flask-restful
flask-sqlalchemy
flask-cors
bcrypt
flask-mail
flask-caching
redis
celery